+++ mlflow/server/js/src/common/utils/FetchUtils.ts
@@ -8,8 +8,9 @@
 import cookie from 'cookie';
 import JsonBigInt from 'json-bigint';
 import yaml from 'js-yaml';
-import { isNil, pickBy } from 'lodash';
+import { pickBy } from 'lodash';
 import { ErrorWrapper } from './ErrorWrapper';
+import { matchPredefinedError } from '@databricks/web-shared/errors';
 
 export const HTTPMethods = {
   GET: 'GET',
@@ -440,4 +441,35 @@ export const deleteYaml = (props: any) => {
     body: generateJsonBody(data),
     success: yamlResponseParser,
   });
+};
+
+function serializeRequestBody(payload: any | FormData | Blob) {
+  if (payload === undefined) {
+    return undefined;
+  }
+  return typeof payload === 'string' || payload instanceof FormData || payload instanceof Blob
+    ? payload
+    : JSON.stringify(payload);
+}
+
+export const fetchAPI = async (url: string, method: 'POST' | 'GET' | 'PATCH' | 'DELETE' = 'GET', body?: any) => {
+  const response = await fetch(url, {
+    method,
+    body: serializeRequestBody(body),
+    headers: body ? { 'Content-Type': 'application/json' } : {},
+  });
+  if (!response.ok) {
+    const predefinedError = matchPredefinedError(response);
+    if (predefinedError) {
+      try {
+        // Attempt to use message from the response
+        const message = (await response.json()).message;
+        predefinedError.message = message ?? predefinedError.message;
+      } catch {
+        // If the message can't be parsed, use default one
+      }
+      throw predefinedError;
+    }
+  }
+  return response.json();
 };
