+++ mlflow/server/js/src/experiment-tracking/hooks/logged-models/useGetLoggedModelQuery.tsx
@@ -1,7 +1,6 @@
 import { type QueryFunctionContext, useQuery } from '@mlflow/mlflow/src/common/utils/reactQueryHooks';
 import type { LoggedModelProto } from '../../types';
-import { loggedModelsDataRequest } from './request.utils';
-import { getAjaxUrl } from '@mlflow/mlflow/src/common/utils/FetchUtils';
+import { fetchAPI, getAjaxUrl } from '@mlflow/mlflow/src/common/utils/FetchUtils';
 
 type UseGetLoggedModelQueryResponseType = {
   model: LoggedModelProto;
@@ -14,7 +13,7 @@ const getQueryKey = (loggedModelId: string): UseGetLoggedModelQueryKey => ['GET_
 const queryFn = async ({
   queryKey: [, loggedModelId],
 }: QueryFunctionContext<UseGetLoggedModelQueryKey>): Promise<UseGetLoggedModelQueryResponseType> =>
-  loggedModelsDataRequest(getAjaxUrl(`ajax-api/2.0/mlflow/logged-models/${loggedModelId}`), 'GET');
+  fetchAPI(getAjaxUrl(`ajax-api/2.0/mlflow/logged-models/${loggedModelId}`), 'GET');
 
 /**
  * Retrieve logged model from API based on its ID
@@ -58,7 +57,7 @@ export const asyncGetLoggedModel = async (
   failSilently = false,
 ): Promise<UseGetLoggedModelQueryResponseType | undefined> => {
   try {
-    const data = await loggedModelsDataRequest(getAjaxUrl(`ajax-api/2.0/mlflow/logged-models/${loggedModelId}`), 'GET');
+    const data = await fetchAPI(getAjaxUrl(`ajax-api/2.0/mlflow/logged-models/${loggedModelId}`), 'GET');
     return data;
   } catch (error) {
     if (failSilently) {
