+++ mlflow/server/js/src/experiment-tracking/hooks/logged-models/usePatchLoggedModelsTags.tsx
@@ -1,7 +1,6 @@
 import { useMutation } from '@mlflow/mlflow/src/common/utils/reactQueryHooks';
 import { entries } from 'lodash';
-import { loggedModelsDataRequest } from './request.utils';
-import { getAjaxUrl } from '@mlflow/mlflow/src/common/utils/FetchUtils';
+import { fetchAPI, getAjaxUrl } from '@mlflow/mlflow/src/common/utils/FetchUtils';
 
 export const usePatchLoggedModelsTags = ({ loggedModelId }: { loggedModelId?: string }) => {
   const { isLoading, error, mutateAsync } = useMutation<unknown, Error, Record<string, string>>({
@@ -10,11 +9,7 @@ export const usePatchLoggedModelsTags = ({ loggedModelId }: { loggedModelId?: st
         tags: entries(variables).map(([key, value]) => ({ key, value })),
       };
 
-      return loggedModelsDataRequest(
-        getAjaxUrl(`ajax-api/2.0/mlflow/logged-models/${loggedModelId}/tags`),
-        'PATCH',
-        requestBody,
-      );
+      return fetchAPI(getAjaxUrl(`ajax-api/2.0/mlflow/logged-models/${loggedModelId}/tags`), 'PATCH', requestBody);
     },
   });
