+++ mlflow/server/js/src/experiment-tracking/pages/experiment-evaluation-datasets/hooks/useSearchEvaluationDatasets.tsx
@@ -1,6 +1,6 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { EvaluationDataset } from '../types';
-import { postJson } from '@mlflow/mlflow/src/common/utils/FetchUtils';
+import { fetchAPI, getAjaxUrl } from '@mlflow/mlflow/src/common/utils/FetchUtils';
 import { useMemo } from 'react';
 import { SEARCH_EVALUATION_DATASETS_QUERY_KEY } from '../constants';
 
@@ -35,10 +35,11 @@ export const useSearchEvaluationDatasets = ({
         page_token: pageParam,
       };
 
-      return (await postJson({
-        relativeUrl: 'ajax-api/3.0/mlflow/datasets/search',
-        data: requestBody,
-      })) as SearchEvaluationDatasetsResponse;
+      return (await fetchAPI(
+        getAjaxUrl('ajax-api/3.0/mlflow/datasets/search'),
+        'POST',
+        requestBody,
+      )) as SearchEvaluationDatasetsResponse;
     },
     cacheTime: 0,
     refetchOnWindowFocus: false,
